api/requirements.txt
Flask
discord.py
gunicorn
flask-cors
